backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
supabase
python-telegram-bot
Pillow
pydantic
pydantic-settings
pydantic[email]
aiohttp
requests
APScheduler
httpx
openai
pytesseract
python-multipart
google-generativeai
opik
litellm